Also support CouchDB2-style string update sequence.
_What_

Cloudant (and the soon to be released CouchDB2) has a different format for the update sequence
compared with CouchDB1.X. In CouchDB1, this was an integer, but in CouchDB2 and its derivatives,
this is an opaque string:

{"results":[
{"seq":
    "1-g1AAAAEceJzLYWBgYMlgTmGQS0lKzi9KdUhJMtIryC8qSS3SS87JL01JzCvRy0styQGqY0pkSLL___
    9_ViIDqg5DnDqSHIBkUj1IUwZzImMukMdubmSaZmFiStAAYu3IYwGSDA1ACmjNfhIcB9F4AKIR6ECg_8AOTDY1sr
    RINcoCAChmW4s",
"id":
    "4855590576c378f78af56f33222c50d1",
"changes":[{
    "rev":"1-b3ef851824b8b32c215e651c32fb0be0"
}]}, ...etc

This makes go-couchdb also compatible with Cloudant and CouchDB2.

In order to pull out `Seq` with the appropriate type, use:

if feed, err := c.DB("db").Changes(nil); err == nil {
    feed.Next()
    if seq, ok := feed.Seq.(int64); ok {
        // seq is int64...
    }
}

Also added a Post().

_How_

* The type of the `Seq` in the ChangesFeed struct was changed from int64 to interface{} so that it can
represent both flavours.
* `func (f *ChangesFeed) contParser(r io.Reader)` was changed to reflect this type change, and also extended
with `Changes` and `Doc` fields so that it also works with `include_docs=true`.
* `func (f *ChangesFeed) pollParser(scan *scanner)` was extended to cater for both types of `last_seq`.

_Testing_

Tests in couchdb2feed_test.go to test the string seq. Added a test for Post().